Fix overflowing indices with files > 1gb
Presently, compressing a file with size over 1024 MB causes attempts to index past available buffers in the BT4 matchfinders. This was occuring due to indexing with a 32-bit value generated by bit shifting `-1<<32`, which due to overflow, creates an unexpectedly large value. As a quick fix, replacing all 32-bit variables with 64-bit values in all internal operarations fixed this issue. In the future, it will be best to look into this issue more, restoring usage of 32-bit values where it is both optimal and non-limiting.

+++ src/decoder_alloc.rs
@@ -31,7 +31,7 @@ impl DerefMut for LZMADecoder {
 }
 
 impl LZMADecoder {
-    pub fn new(lc: u32, lp: u32, pb: u32) -> Self {
+    pub fn new(lc: u64, lp: u64, pb: u64) -> Self {
         let mut literal_decoder = LiteralDecoder::new(lc, lp);
         literal_decoder.reset();
         let match_len_decoder = {
@@ -84,7 +84,7 @@ impl LZMADecoder {
             }
         }
         while lz.has_space() {
-            let pos_state = lz.get_pos() as u32 & self.pos_mask;
+            let pos_state = lz.get_pos() as u64 & self.pos_mask;
             let i = self.coder.state.get() as usize;
             let probs = &mut self.is_match[i];
             let bit = rc.decode_bit(&mut probs[pos_state as usize])?;
@@ -120,9 +120,9 @@ impl LZMADecoder {
 
     fn decode_match<R: RangeSource>(
         &mut self,
-        pos_state: u32,
+        pos_state: u64,
         rc: &mut RangeDecoder<R>,
-    ) -> crate::io::read_exact_result!(R, u32) {
+    ) -> crate::io::read_exact_result!(R, u64) {
         self.coder.state.update_match();
         self.reps[3] = self.reps[2];
         self.reps[2] = self.reps[1];
@@ -131,16 +131,16 @@ impl LZMADecoder {
         let len = self.match_len_decoder.decode(pos_state as _, rc)?;
         let dist_slot = rc.decode_bit_tree(&mut self.dist_slots[coder_get_dict_size(len as _)])?;
 
-        if dist_slot < DIST_MODEL_START as i32 {
+        if dist_slot < DIST_MODEL_START as i64 {
             self.reps[0] = dist_slot as _;
         } else {
             let limit = (dist_slot >> 1) - 1;
             self.reps[0] = (2 | (dist_slot & 1)) << limit;
-            if dist_slot < DIST_MODEL_END as i32 {
-                let probs = self.get_dist_special((dist_slot - DIST_MODEL_START as i32) as usize);
+            if dist_slot < DIST_MODEL_END as i64 {
+                let probs = self.get_dist_special((dist_slot - DIST_MODEL_START as i64) as usize);
                 self.reps[0] |= rc.decode_reverse_bit_tree(probs)?;
             } else {
-                let r0 = rc.decode_direct_bits(limit as u32 - ALIGN_BITS as u32)? << ALIGN_BITS;
+                let r0 = rc.decode_direct_bits(limit as u64 - ALIGN_BITS as u64)? << ALIGN_BITS;
                 self.reps[0] |= r0;
                 self.reps[0] |= rc.decode_reverse_bit_tree(&mut self.dist_align)?;
             }
@@ -151,9 +151,9 @@ impl LZMADecoder {
 
     fn decode_rep_match<R: RangeSource>(
         &mut self,
-        pos_state: u32,
+        pos_state: u64,
         rc: &mut RangeDecoder<R>,
-    ) -> crate::io::read_exact_result!(R, u32) {
+    ) -> crate::io::read_exact_result!(R, u64) {
         let index = self.coder.state.get() as usize;
         if rc.decode_bit(&mut self.is_rep0[index])? == 0 {
             let index: usize = self.coder.state.get() as usize;
@@ -182,7 +182,7 @@ impl LZMADecoder {
         self.coder.state.update_long_rep();
         self.rep_len_decoder
             .decode(pos_state as _, rc)
-            .map(|i| i as u32)
+            .map(|i| i as u64)
     }
 }
 pub struct LiteralDecoder {
@@ -191,7 +191,7 @@ pub struct LiteralDecoder {
 }
 
 impl LiteralDecoder {
-    fn new(lc: u32, lp: u32) -> Self {
+    fn new(lc: u64, lp: u64) -> Self {
         let coder = LiteralCoder::new(lc, lp);
         let sub_decoders = vec![LiteralSubdecoder::new(); (1 << (lc + lp)) as _];
 
@@ -238,11 +238,11 @@ impl LiteralSubdecoder {
         lz: &mut LZDecoder,
         rc: &mut RangeDecoder<R>,
     ) -> crate::io::read_exact_result!(R, ()) {
-        let mut symbol: u32 = 1;
+        let mut symbol: u64 = 1;
         let liter = coder.state.is_literal();
         if liter {
             loop {
-                let b = rc.decode_bit(&mut self.coder.probs[symbol as usize])? as u32;
+                let b = rc.decode_bit(&mut self.coder.probs[symbol as usize])? as u64;
                 symbol = (symbol << 1) | b;
                 if symbol >= 0x100 {
                     break;
@@ -250,7 +250,7 @@ impl LiteralSubdecoder {
             }
         } else {
             let r = coder.reps[0];
-            let mut match_byte = lz.get_byte(r as usize) as u32;
+            let mut match_byte = lz.get_byte(r as usize) as u64;
             let mut offset = 0x100;
             let mut match_bit;
             let mut bit;
@@ -260,9 +260,9 @@ impl LiteralSubdecoder {
                 match_bit = match_byte & offset;
                 bit = rc
                     .decode_bit(&mut self.coder.probs[(offset + match_bit + symbol) as usize])?
-                    as u32;
+                    as u64;
                 symbol = (symbol << 1) | bit;
-                offset &= (0u32.wrapping_sub(bit)) ^ !match_bit;
+                offset &= (0u64.wrapping_sub(bit)) ^ !match_bit;
                 if symbol >= 0x100 {
                     break;
                 }
@@ -279,7 +279,7 @@ impl LengthCoder {
         &mut self,
         pos_state: usize,
         rc: &mut RangeDecoder<R>,
-    ) -> crate::io::read_exact_result!(R, i32) {
+    ) -> crate::io::read_exact_result!(R, i64) {
         if rc.decode_bit(&mut self.choice[0])? == 0 {
             return Ok(rc
                 .decode_bit_tree(&mut self.low[pos_state])?
